refactor: move validators to domain layer
