+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -8,7 +8,7 @@ import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.menu.repository.MenuRepository;
-import kitchenpos.menu.validator.MenuValidator;
+import kitchenpos.menu.domain.MenuValidator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
