+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -19,7 +19,7 @@ import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.menu.repository.MenuRepository;
-import kitchenpos.menu.validator.MenuValidator;
+import kitchenpos.menu.domain.MenuValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
