+++ src/test/java/kitchenpos/table/validator/OrderCreationValidatorTest.java
@@ -10,6 +10,7 @@ import java.util.Optional;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItems;
 import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.table.domain.OrderCreationValidator;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.repository.OrderTableRepository;
 import org.junit.jupiter.api.DisplayName;
